+++ deltafq/live/__init__.py
@@ -8,9 +8,11 @@ from .gateways import DataGateway, TradeGateway
 from ..adapters.data import YFinanceDataGateway
 from ..adapters.trade import PaperTradeGateway
 from .gateway_registry import DATA_GATEWAYS, TRADE_GATEWAYS, create_data_gateway, create_trade_gateway
+from .engine import LiveEngine
 
 __all__ = [
     "EventEngine",
+    "LiveEngine",
     "TickData",
     "OrderRequest",
     "DataGateway",